move UpdateParams Methods in independent files: mint, slashing, staking modules
