+++ modules/mint/handle_block.go
@@ -3,10 +3,7 @@ package mint
 import (
 	juno "github.com/forbole/juno/v2/types"
 
-	"github.com/rs/zerolog/log"
 	tmctypes "github.com/tendermint/tendermint/rpc/core/types"
-
-	"github.com/forbole/bdjuno/v2/types"
 )
 
 // HandleBlock implements modules.BlockModule
@@ -14,26 +11,4 @@ func (m *Module) HandleBlock(
 	block *tmctypes.ResultBlock, _ *tmctypes.ResultBlockResults, _ []*juno.Tx, _ *tmctypes.ResultValidators,
 ) error {
 	return nil
-}
-
-// UpdateParams gets the updated params and stores them inside the database
-func (m *Module) UpdateParams(height int64) {
-	log.Debug().Str("module", "mint").Int64("height", height).
-		Msg("updating params")
-
-	params, err := m.source.Params(height)
-	if err != nil {
-		log.Error().Str("module", "mint").Err(err).
-			Int64("height", height).
-			Msg("error while getting params")
-		return
-	}
-
-	err = m.db.SaveMintParams(types.NewMintParams(params, height))
-	if err != nil {
-		log.Error().Str("module", "mint").Err(err).
-			Int64("height", height).
-			Msg("error while saving params")
-		return
-	}
 }
