+++ modules/staking/handle_block.go
@@ -43,28 +43,6 @@ func (m *Module) HandleBlock(
 	return nil
 }
 
-// UpdateParams gets the updated params and stores them inside the database
-func (m *Module) UpdateParams(height int64) {
-	log.Debug().Str("module", "staking").Int64("height", height).
-		Msg("updating params")
-
-	params, err := m.source.GetParams(height)
-	if err != nil {
-		log.Error().Str("module", "staking").Err(err).
-			Int64("height", height).
-			Msg("error while getting params")
-		return
-	}
-
-	err = m.db.SaveStakingParams(types.NewStakingParams(params, height))
-	if err != nil {
-		log.Error().Str("module", "staking").Err(err).
-			Int64("height", height).
-			Msg("error while saving params")
-		return
-	}
-}
-
 // updateValidatorsStatus updates all validators' statuses
 func (m *Module) updateValidatorsStatus(height int64, validators []stakingtypes.Validator) {
 	log.Debug().Str("module", "staking").Int64("height", height).
